+++ src/mol-gl/shader/postprocessing.frag.ts
@@ -1,8 +1,9 @@
 /**
- * Copyright (c) 2019-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2019-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  * @author Áron Samuel Kovács <aron.kovacs@mail.muni.cz>
+ * @author Gianluca Tomasello <giagitom@gmail.com>
  */
 
 export const postprocessing_frag = `
@@ -11,8 +12,11 @@ precision highp int;
 precision highp sampler2D;
 
 uniform sampler2D tSsaoDepth;
+uniform sampler2D tSsaoDepthTransparent;
 uniform sampler2D tColor;
+uniform sampler2D tTransparentColor;
 uniform sampler2D tDepthOpaque;
+uniform sampler2D tDepthTransparent;
 uniform sampler2D tShadows;
 uniform sampler2D tOutlines;
 uniform vec2 uTexSize;
@@ -45,19 +49,24 @@ float getDepthOpaque(const in vec2 coords) {
     #endif
 }
 
-bool isBackground(const in float depth) {
-    return depth == 1.0;
+float getDepthTransparent(const in vec2 coords) {
+    return unpackRGBAToDepthWithAlpha(texture2D(tDepthTransparent, coords)).x;
 }
 
-float getOutline(const in vec2 coords, const in float opaqueDepth, out float closestTexel) {
-    float backgroundViewZ = 2.0 * uFar;
+bool isBackground(const in float depth) {
+    return depth > 0.999; // handle depth packing precision issues
+}
+
+int squaredOutlineScale = dOutlineScale * dOutlineScale;
+float getOutline(const in vec2 coords, const in float opaqueDepth, const in float transparentDepth, out float closestTexel, out float isTransparent) {
     vec2 invTexSize = 1.0 / uTexSize;
 
     float outline = 1.0;
     closestTexel = 1.0;
+    isTransparent = 0.0;
     for (int y = -dOutlineScale; y <= dOutlineScale; y++) {
         for (int x = -dOutlineScale; x <= dOutlineScale; x++) {
-            if (x * x + y * y > dOutlineScale * dOutlineScale) {
+            if (x * x + y * y > squaredOutlineScale) {
                 continue;
             }
 
@@ -70,10 +79,11 @@ float getOutline(const in vec2 coords, const in float opaqueDepth, out float clo
             if (sampleOutline == 0.0 && sampleOutlineDepth < closestTexel) {
                 outline = 0.0;
                 closestTexel = sampleOutlineDepth;
+                isTransparent = sampleOutlineCombined.a;
             }
         }
     }
-    return closestTexel < opaqueDepth ? outline : 1.0;
+    return isTransparent == 0.0 ? outline : (closestTexel > opaqueDepth && closestTexel < transparentDepth) ? 1.0 : outline;
 }
 
 float getSsao(vec2 coords) {
@@ -87,31 +97,60 @@ float getSsao(vec2 coords) {
     return 1.0;
 }
 
+float getSsaoTransparent(vec2 coords) {
+    float rawSsao = unpackRGToUnitInterval(texture2D(tSsaoDepthTransparent, coords).xy);
+    if (rawSsao > 0.999) {
+        return 1.0;
+    } else if (rawSsao > 0.001) {
+        return rawSsao;
+    }
+    // treat values close to 0.0 as errors and return no occlusion
+    return 1.0;
+}
+
 void main(void) {
     vec2 coords = gl_FragCoord.xy / uTexSize;
     vec4 color = texture2D(tColor, coords);
 
-    float viewDist;
-    float fogFactor;
     float opaqueDepth = getDepthOpaque(coords);
+    float transparentDepth = 1.0;
+    #ifdef dBlendTransparency
+        bool blendTransparency = true;
+        vec4 transparentColor = texture2D(tTransparentColor, coords);
 
-    #ifdef dOcclusionEnable
-        if (!isBackground(opaqueDepth)) {
-            viewDist = abs(getViewZ(opaqueDepth));
-            fogFactor = smoothstep(uFogNear, uFogFar, viewDist);
+        #if defined(dOutlineEnable) || defined(dOcclusionEnable) && defined(dOcclusionIncludeTransparency)
+            transparentDepth = getDepthTransparent(coords);
+        #endif
+    #endif
+
+    #if defined(dOcclusionEnable) || defined(dShadowEnable)
+        bool isOpaqueBackground = isBackground(opaqueDepth);
+        float viewDist = abs(getViewZ(opaqueDepth));
+        float fogFactor = smoothstep(uFogNear, uFogFar, viewDist);
+    #endif
+
+    #if defined(dOcclusionEnable)
+        if (!isOpaqueBackground) {
             float occlusionFactor = getSsao(coords + uOcclusionOffset);
+
             if (!uTransparentBackground) {
                 color.rgb = mix(mix(uOcclusionColor, uFogColor, fogFactor), color.rgb, occlusionFactor);
             } else {
                 color.rgb = mix(uOcclusionColor * (1.0 - fogFactor), color.rgb, occlusionFactor);
             }
         }
+        #if defined(dBlendTransparency) && defined(dOcclusionIncludeTransparency)
+            if (!isBackground(transparentDepth)) {
+                float viewDist = abs(getViewZ(transparentDepth));
+                float fogFactor = smoothstep(uFogNear, uFogFar, viewDist);
+                float occlusionFactor = getSsaoTransparent(coords + uOcclusionOffset);
+                transparentColor.rgb = mix(uOcclusionColor * (1.0 - fogFactor), transparentColor.rgb, occlusionFactor);
+            }
+        #endif
     #endif
 
     #ifdef dShadowEnable
-        if (!isBackground(opaqueDepth)) {
-            viewDist = abs(getViewZ(opaqueDepth));
-            fogFactor = smoothstep(uFogNear, uFogFar, viewDist);
+        if (!isOpaqueBackground) {
             vec4 shadow = texture2D(tShadows, coords);
             if (!uTransparentBackground) {
                 color.rgb = mix(mix(vec3(0), uFogColor, fogFactor), color.rgb, shadow.a);
@@ -124,15 +163,31 @@ void main(void) {
     // outline needs to be handled after occlusion and shadow to keep them clean
     #ifdef dOutlineEnable
         float closestTexel;
-        float outline = getOutline(coords, opaqueDepth, closestTexel);
+        float isTransparentOutline;
+        float outline = getOutline(coords, opaqueDepth, transparentDepth, closestTexel, isTransparentOutline);
         if (outline == 0.0) {
-            viewDist = abs(getViewZ(closestTexel));
-            fogFactor = smoothstep(uFogNear, uFogFar, viewDist);
+            float viewDist = abs(getViewZ(closestTexel));
+            float fogFactor = smoothstep(uFogNear, uFogFar, viewDist);
             if (!uTransparentBackground) {
-                color.rgb = mix(uOutlineColor, uFogColor, fogFactor);
+                    color.rgb = mix(uOutlineColor, uFogColor, fogFactor);
             } else {
                 color.a = 1.0 - fogFactor;
-                color.rgb = mix(uOutlineColor, color.rgb, fogFactor);
+                color.rgb = mix(uOutlineColor, vec3(0.0), fogFactor);
+            }
+            #ifdef dBlendTransparency
+                if (isTransparentOutline == 1.0 || transparentDepth > closestTexel) {
+                    blendTransparency = false;
+                }
+            #endif
+        }
+    #endif
+
+    #ifdef dBlendTransparency
+        if (blendTransparency) {
+            float alpha = transparentColor.a;
+            if (alpha != 0.0) {
+                // blending
+                color = transparentColor + color * (1.0 - alpha);
             }
         }
     #endif