+++ src/mol-plugin-state/transforms/model.ts
@@ -1,9 +1,10 @@
 /**
- * Copyright (c) 2018-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  * @author Adam Midlik <midlik@gmail.com>
+ * @author Ludovic Autin <ludovic.autin@gmail.com>
  */
 
 import { parseDcd } from '../../mol-io/reader/dcd/parser';
@@ -40,6 +41,11 @@ import { parseXtc } from '../../mol-io/reader/xtc/parser';
 import { coordinatesFromXtc } from '../../mol-model-formats/structure/xtc';
 import { parseXyz } from '../../mol-io/reader/xyz/parser';
 import { trajectoryFromXyz } from '../../mol-model-formats/structure/xyz';
+import { UnitStyles } from '../../mol-io/reader/lammps/schema';
+import { parseLammpsData } from '../../mol-io/reader/lammps/data/parser';
+import { trajectoryFromLammpsData } from '../../mol-model-formats/structure/lammps-data';
+import { parseLammpsTrajectory } from '../../mol-io/reader/lammps/traj/parser';
+import { coordinatesFromLammpsTrajectory, trajectoryFromLammpsTrajectory } from '../../mol-model-formats/structure/lammps-trajectory';
 import { parseSdf } from '../../mol-io/reader/sdf/parser';
 import { trajectoryFromSdf } from '../../mol-model-formats/structure/sdf';
 import { assertUnreachable } from '../../mol-util/type-helpers';
@@ -54,6 +60,7 @@ export { CoordinatesFromDcd };
 export { CoordinatesFromXtc };
 export { CoordinatesFromTrr };
 export { CoordinatesFromNctraj };
+export { CoordinatesFromLammpstraj };
 export { TopologyFromPsf };
 export { TopologyFromPrmtop };
 export { TopologyFromTop };
@@ -63,6 +70,8 @@ export { TrajectoryFromMmCif };
 export { TrajectoryFromPDB };
 export { TrajectoryFromGRO };
 export { TrajectoryFromXYZ };
+export { TrajectoryFromLammpsData };
+export { TrajectoryFromLammpsTrajData };
 export { TrajectoryFromMOL };
 export { TrajectoryFromSDF };
 export { TrajectoryFromMOL2 };
@@ -145,6 +154,23 @@ const CoordinatesFromNctraj = PluginStateTransform.BuiltIn({
             const parsed = await parseNctraj(a.data).runInContext(ctx);
             if (parsed.isError) throw new Error(parsed.message);
             const coordinates = await coordinatesFromNctraj(parsed.result).runInContext(ctx);
+            return new SO.Molecule.Coordinates(coordinates, { label: a.label, description: 'Coordinates' });
+        });
+    }
+});
+
+type CoordinatesFromLammpstraj = typeof CoordinatesFromLammpstraj
+const CoordinatesFromLammpstraj = PluginStateTransform.BuiltIn({
+    name: 'coordinates-from-lammpstraj',
+    display: { name: 'Parse LAMMPSTRAJ', description: 'Parse LAMMPSTRAJ data.' },
+    from: [SO.Data.String],
+    to: SO.Molecule.Coordinates
+})({
+    apply({ a }) {
+        return Task.create('Parse LAMMPSTRAJ', async ctx => {
+            const parsed = await parseLammpsTrajectory(a.data).runInContext(ctx);
+            if (parsed.isError) throw new Error(parsed.message);
+            const coordinates = await coordinatesFromLammpsTrajectory(parsed.result).runInContext(ctx);
             return new SO.Molecule.Coordinates(coordinates, { label: a.label, description: 'Coordinates' });
         });
     }
@@ -374,6 +400,49 @@ const TrajectoryFromXYZ = PluginStateTransform.BuiltIn({
         });
     }
 });
+
+type TrajectoryFromLammpsData = typeof TrajectoryFromLammpsData
+const TrajectoryFromLammpsData = PluginStateTransform.BuiltIn({
+    name: 'trajectory-from-lammps-data',
+    display: { name: 'Parse Lammps Data', description: 'Parse Lammps Data from string and create trajectory.' },
+    from: [SO.Data.String],
+    to: SO.Molecule.Trajectory,
+    params: {
+        unitsStyle: PD.Select('real', PD.arrayToOptions(UnitStyles)),
+    }
+})({
+    apply({ a, params }) {
+        return Task.create('Parse Lammps Data', async ctx => {
+            const parsed = await parseLammpsData(a.data).runInContext(ctx);
+            if (parsed.isError) throw new Error(parsed.message);
+            const models = await trajectoryFromLammpsData(parsed.result, params.unitsStyle).runInContext(ctx);
+            const props = trajectoryProps(models);
+            return new SO.Molecule.Trajectory(models, props);
+        });
+    }
+});
+
+type TrajectoryFromLammpsTrajData = typeof TrajectoryFromLammpsTrajData
+const TrajectoryFromLammpsTrajData = PluginStateTransform.BuiltIn({
+    name: 'trajectory-from-lammps-traj-data',
+    display: { name: 'Parse Lammps traj Data', description: 'Parse Lammps Traj Data string and create trajectory.' },
+    from: [SO.Data.String],
+    to: SO.Molecule.Trajectory,
+    params: {
+        unitsStyle: PD.Select('real', PD.arrayToOptions(UnitStyles)),
+    }
+})({
+    apply({ a, params }) {
+        return Task.create('Parse Lammps Data', async ctx => {
+            const parsed = await parseLammpsTrajectory(a.data).runInContext(ctx);
+            if (parsed.isError) throw new Error(parsed.message);
+            const models = await trajectoryFromLammpsTrajectory(parsed.result, params.unitsStyle).runInContext(ctx);
+            const props = trajectoryProps(models);
+            return new SO.Molecule.Trajectory(models, props);
+        });
+    }
+});
+
 
 type TrajectoryFromMOL = typeof TrajectoryFromMOL
 const TrajectoryFromMOL = PluginStateTransform.BuiltIn({