+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- Add SSAO support for transparent geometry
+- Fix SSAO color not updating
+- Improve blending of overlapping outlines from transparent & opaque geometries
 - Default to `blended` transparency on iOS due to `wboit` not being supported.
 - Fix direct-volume with fog off (and on with `dpoit`) and transparent background on (#1286)
 - Fix missing pre-multiplied alpha for `blended` & `wboit` with no fog (#1284)
@@ -17,10 +20,14 @@ Note that since we don't clearly distinguish between a public and private interf
 - MolViewSpec: MVP Support for geometrical primitives (mesh, lines, line, label, distance measurement)
 - Mesoscale Explorer: Add support for 4-character PDB IDs (e.g., 8ZZC) in PDB-Dev loader
 - Fix Sequence View in Safari 18
+<<<<<<< HEAD
 - ModelArchive QualityAssessment extension:
   - Add support for ma_qa_metric_local_pairwise mmCIF category
   - Add PAE plot component
 - Added new AlphaFoldDB-PAE example
+=======
+- Add support for LAMMPS data and dump formats
+>>>>>>> 84e292b3e28d424f179bfd8851617216f778fee7
 
 ## [v4.7.1] - 2024-09-30